update staggered packages
