+++ pkgs/jobs/job_manager.go
@@ -1,4 +1,4 @@
-package perform
+package jobs
 
 import (
 	"strings"
