+++ pkgs/jobs/jobs_test_jobs.go
@@ -1,4 +1,4 @@
-package perform
+package jobs
 
 import (
 	"encoding/hex"
@@ -6,7 +6,7 @@ import (
 	"strconv"
 
 	"github.com/eris-ltd/eris-cli/definitions"
-	"github.com/eris-ltd/eris-cli/ebi"
+	"github.com/eris-ltd/eris-cli/pkgs/abi"
 	"github.com/eris-ltd/eris-cli/log"
 	"github.com/eris-ltd/eris-cli/util"
 
@@ -39,10 +39,10 @@ func QueryContractJob(query *definitions.QueryContract, do *definitions.Do) (str
 	var data string
 	var packedBytes []byte
 	if query.ABI == "" {
-		packedBytes, err = ebi.ReadAbiFormulateCall(query.Destination, query.Function, queryDataArray, do)
+		packedBytes, err = abi.ReadAbiFormulateCall(query.Destination, query.Function, queryDataArray, do)
 		data = hex.EncodeToString(packedBytes)
 	} else {
-		packedBytes, err = ebi.ReadAbiFormulateCall(query.ABI, query.Function, queryDataArray, do)
+		packedBytes, err = abi.ReadAbiFormulateCall(query.ABI, query.Function, queryDataArray, do)
 		data = hex.EncodeToString(packedBytes)
 	}
 	if err != nil {
@@ -65,10 +65,10 @@ func QueryContractJob(query *definitions.QueryContract, do *definitions.Do) (str
 	log.WithField("res", result).Debug("Decoding Raw Result")
 	if query.ABI == "" {
 		log.WithField("abi", query.Destination).Debug()
-		query.Variables, err = ebi.ReadAndDecodeContractReturn(query.Destination, query.Function, result, do)
+		query.Variables, err = abi.ReadAndDecodeContractReturn(query.Destination, query.Function, result, do)
 	} else {
 		log.WithField("abi", query.ABI).Debug()
-		query.Variables, err = ebi.ReadAndDecodeContractReturn(query.ABI, query.Function, result, do)
+		query.Variables, err = abi.ReadAndDecodeContractReturn(query.ABI, query.Function, result, do)
 	}
 	if err != nil {
 		return "", make([]*definitions.Variable, 0), err
